requirements.txt
weaviate-client
pandas
requests
tqdm